[Evo][ContribFolder] fix #22006 #21980 #21980 #19572
