+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "entcore",
-  "version": "3.3.0-dev.4",
+  "version": "3.3.0-dev.5",
   "description": "",
   "main": "gulpfile.js",
   "types": "./types/src/ts/entcore.d.ts",
